separate the  Music Slider component
