+++ src/pages/HomePage/HomePage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {useSearchQuery} from "../../hooks/useSearchQuery"
-import NewReleases from "../../common/MusicCard/NewReleases";
+import NewReleases from '../../common/MusicCard/NewReleases/NewReleases.jsx';
 
 
 const HomePage = () => {
@@ -12,7 +12,7 @@ const HomePage = () => {
     <div>
       <h1>Homepage</h1>
 
-      {/* 인기 앨범 슬라이드 추가 */}
+      {/* 신규앨범 슬라이드 추가 */}
       <NewReleases />
     </div>
   )
